Добавил RRT
1. Сделал rrt
2. Добавл команду rrt
3. Добавил общий генератор случайных чисел с режимами в конфиг
4. Немного рефакторинга
5. Добавил непрерывные методы
6. Улучшил безопасность config
7. Починил визуализацию gnuplot + цветной bmp
8. Убрал ненужные режимы bin
9. Протестировал корректность программы
10. Добавил строгие флаги компилятору и линковщику. Программу исправил, чтобы компилировалась
11. Добавил визуализацию RRT
12. Обновил датасеты в исследованиях, после исправлений в коде
13. Новое исследование RRT

+++ src/algorithms/path/common/path_metrics.hpp
@@ -14,7 +14,7 @@ struct PathMetrics {
     size_t pixelLength = 0;                         // Длина пути в пикселях (Bresenham)
     bool pathFound = false;                         // Найден ли путь
     double minObstacleDistance = 0;                 // Минимальное Евклидово расстояние до препятсвия
-    size_t minObstacleDistancePixel = 0;            // Минимальное пиксельное расстояние до препятсвия
+    int minObstacleDistancePixel = 0;               // Минимальное пиксельное расстояние до препятсвия
     double maxSideAngle = 0;                        // Максимальный наклон вбок
     double maxUpDownAngle = 0;                      // Максимальный наклон вперед/назад
 };
